Refactored dates to use CbetDatePicker.
